be/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "numpy>=2.4.3",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit (security)
    "RUF", # ruff-specific rules
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # pytest uses assert statements

[tool.ruff.lint.isort]
known-first-party = ["routers", "services", "store", "models"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.pyright]
pythonVersion = "3.14"
venvPath = "."
venv = ".venv"
